requirements.txt
discord.py
aiosqlite
Pillow
aiohttp
flask